refactor: move business logic from CardController to CardService
